fix: 페이지 오프셋 long→int 캐스팅 오버플로우 방어
(int) pageable.getOffset() 대신 Math.toIntExact() 사용으로
정수 범위 초과 시 ArithmeticException으로 명시적 실패 처리

+++ SSAju/src/main/java/ssafy/SSAju/admin/repository/AdminFeedbackQueryRepository.java
@@ -39,7 +39,7 @@ public class AdminFeedbackQueryRepository {
             countQuery.setParameter("type", feedbackType);
         }
 
-        query.setFirstResult((int) pageable.getOffset());
+        query.setFirstResult(Math.toIntExact(pageable.getOffset()));
         query.setMaxResults(pageable.getPageSize());
 
         List<FeedbackListDTO> content = query.getResultList().stream()
